requirements.txt
# requirements.txt
# Install these dependencies using: pip install -r requirements.txt

Flask
python-dotenv
ldap3
gunicorn
